requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
alembic==1.8.0
    # via flask-migrate
certifi==2022.6.15
    # via requests
charset-normalizer==2.0.12
    # via requests
click==7.1.2
    # via
    #   flask
    #   pip-compile-multi
    #   pip-tools
flask==1.1.4
    # via
    #   -r requirements.in
    #   flask-migrate
    #   flask-script
    #   flask-sqlalchemy
flask-migrate==3.1.0
    # via -r requirements.in
flask-script==2.0.6
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via alembic
importlib-resources==5.8.0
    # via alembic
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via
    #   -r requirements.in
    #   flask
mako==1.2.0
    # via alembic
markupsafe==2.0.0
    # via
    #   -r requirements.in
    #   jinja2
    #   mako
    #   wtforms
pep517==0.12.0
    # via pip-tools
pip-compile-multi==2.4.5
    # via -r requirements.in
pip-tools==6.6.2
    # via pip-compile-multi
psycopg2==2.9.3
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.28.0
    # via -r requirements.in
sqlalchemy==1.4.39
    # via
    #   alembic
    #   flask-sqlalchemy
tomli==2.0.1
    # via pep517
toposort==1.7
    # via pip-compile-multi
urllib3==1.26.9
    # via requests
werkzeug==1.0.1
    # via
    #   -r requirements.in
    #   flask
wheel==0.37.1
    # via pip-tools
wtforms==3.0.1
    # via -r requirements.in
zipp==3.8.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
